Added p-model directive verdict
p-model is already inside the framework (as a To be Implemented feature), so feel free to add it to the framework if you feel like your project requires it.

I decided to leave p-model's out of the PHPue framework because even though it's a exciting feature, there is alternatives to replicate the same design flow. I felt that p-model's take the fun out of PHP development, and creates more abstractions, which overcomplicate development again. The main directives PHPue needs is p-if, and p-for, anything else in my opinion, will be forcing developers to lock in to a new way of thinking.

Also, tracking two-way binding especially on forms, breaches privacy, and GDPR concerns. Developers currently are locked into two-way binding methods on other frameworks without a strong hold on how to escape the trap. This way, we can add two-way tracking if we need to, by the use of AJAX and CSR listener tricks.  For when it's required, like search boxes etc.

I updated the Documentation, and fixed some typos.